requirements[1].txt
flask
tensorflow
numpy
pillow
gunicorn
scikit-learn
opencv-python
